Change domain model so that we don't have a formal foreign key relationship between Project/ProjectFunctionality/ProjectRequirement and GrouseUser.
My impression that having these onetomany and manytoone relationships is
overkill and we may take a performance hit because of them. So now
we just use a ownedby field, at the expense of loosing referential
integrity.

Also some tidyup in code and make sure everything that an authorised
user can only touch his/her own things.