Fix syntax error for priority field
In the main XML source for the database 'O (i)' uses in some cases 'O(i)'.
The GUI expects 'O (i)' so 'O(i)' is unrecognised. This results in no
value appearing in the GUI. The same applies to '1(i)' and '2(i)'.

You can see if it is a problem with the following SQL-command:

   SELECT priority FROM requirements group by priority;

In the database the problem can be rectified with:

   update requirements set priority = 'O (i)' where priority = 'O(i)';
   update requirements set priority = '1 (i)' where priority = '1(i)';
   update requirements set priority = '2 (i)' where priority = '2(i)';
   update project_requirements set priority = 'O (i)' where priority = 'O(i)';
   update project_requirements set priority = '1 (i)' where priority = '1(i)';
   update project_requirements set priority = '2 (i)' where priority = '2(i)';

This fixes the problem for future projects.