Fixed refreshing issue
There was an issue where you had to log in each time you refreshed the page, this naturaly is might annoying when you are trying to make a website. This does it so that your browser keeps the auth token aswell as information about who you are and where you were on the website across refreshes.