Fetches projects from server
The client now fetches projects owned by the currently signed in user from the server. With the addition of this it now also fixes many issues related to previous login and out functionality. The logout button now also sends a request to invalidate the token, wich it previously didn`t, and should you be taken to the main-menu without a valid token you will now be sent back to complete the login process