Better support for testing
While trying to develop testing, I stumbled across a number of problems,
in particular, that the @CreatedBy ownedBy had a null value. There is an
anonymousUser in place during testing and the MockitoTest user is being
ignored. I ended up disabling the production ResourceServerConfiguration
with one defined in test and some other minor additions.