Progressbar and unfinished functionalities added to statuspage
There is now a progressbar that shows overall progress on the project, if there are any unfinished requirments that the user has navigated past they will be shown so that they can easely be navigated back to. There are also a bunch of fixes relating to the proccesed tag on all projectfunctionalities, mainly that now when the program detects that all children of a parent are approved, the parent will also be approved. Next up is adding a download button to this page.