Add rel/href to user list for admin role
Note that this means the rel/href will also appear when you,
as admin, retrieve another user. This is marked with --> in
the JSON below.

{
  "username": "user@example.com",
  "accountNonExpired": true,
  "credentialsNonExpired": true,
  "accountNonLocked": true,
  "enabled": true,
  "_links": {
    "self": {
      "href": "http://localhost:9294/grouse/user/user@example.com"
    },
    "project": {
      "href": "http://localhost:9294/grouse/project"
    },
    "template": {
      "href": "http://localhost:9294/grouse/template"
    },
-->    "user": {
-->      "href": "http://localhost:9294/grouse/user"
    }
  }
}

I am not sure if I want this functionality or not. It
is correct that the admin user should see this from a
browseabilty point of view.