Better handling of ProjectFunctionality
ProjectFunctionality was returning all children
(2 layers down) and ProjectRequirements. This
brings a cleaner approach where
project->ProjectFunctionality returns a single
layer and the client must traverse down to lower
levels. The reason for this is that originally
grouse was hardcoded for three layers, but there
could be 2 or 4. The GUI can support 2 or 3 layers.