Fixed the error that broke the sidenav after adding or removing a req
Also ploished up the loading of data so that the console wont spit a ton of error out when you first load the page. Also fixed up the the sidestatusbar so that it now functions properly. Aswell as fixing the endpage and approving functionalities, this reuired that i added a parent datafield to all projectfunctionalities, this might be useful for other things aswell.