Finalise update to Template
Template and project are basically the same so this has been
an attempt to get the two to in "sync". Template will need a
little more work though. There is a difference in how they
are created. A Project is created from a template, while a
template is created from scratch.